requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
requests==2.32.3
setuptools<82
numpy>=1.26.0
Pillow>=10.4.0
face-recognition==1.3.0
face_recognition_models==0.3.0